Initial refactoring of Hadoop package
