+++ src/main/java/ldbc/snb/datagen/hadoop/generator/HadoopPersonActivityGenerator.java
@@ -33,12 +33,19 @@
  You should have received a copy of the GNU General Public License
  along with this program; if not, write to the Free Software
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.*/
-package ldbc.snb.datagen.hadoop;
+package ldbc.snb.datagen.hadoop.generator;
 
 import ldbc.snb.datagen.dictionary.Dictionaries;
 import ldbc.snb.datagen.DatagenParams;
 import ldbc.snb.datagen.LDBCDatagen;
 import ldbc.snb.datagen.generator.generators.PersonActivityGenerator;
+import ldbc.snb.datagen.hadoop.HadoopBlockMapper;
+import ldbc.snb.datagen.hadoop.HadoopBlockPartitioner;
+import ldbc.snb.datagen.hadoop.key.blockkey.BlockKey;
+import ldbc.snb.datagen.hadoop.key.blockkey.BlockKeyComparator;
+import ldbc.snb.datagen.hadoop.key.blockkey.BlockKeyGroupComparator;
+import ldbc.snb.datagen.hadoop.key.TupleKey;
+import ldbc.snb.datagen.hadoop.miscjob.HadoopFileRanker;
 import ldbc.snb.datagen.objects.Knows;
 import ldbc.snb.datagen.objects.Person;
 import ldbc.snb.datagen.serializer.DynamicActivitySerializer;
