+++ src/main/java/ldbc/snb/datagen/hadoop/miscjob/keychanger/InterestKeySetter.java
@@ -33,28 +33,18 @@
  You should have received a copy of the GNU General Public License
  along with this program; if not, write to the Free Software
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.*/
-package ldbc.snb.datagen.hadoop;
+package ldbc.snb.datagen.hadoop.miscjob.keychanger;
 
-import org.apache.hadoop.io.WritableComparable;
-import org.apache.hadoop.io.WritableComparator;
+import ldbc.snb.datagen.hadoop.key.TupleKey;
+import ldbc.snb.datagen.objects.Person;
 
 /**
  * Created by aprat on 11/17/14.
  */
+public class InterestKeySetter implements HadoopFileKeyChanger.KeySetter<TupleKey> {
 
-public class BlockKeyGroupComparator extends WritableComparator {
-
-    protected BlockKeyGroupComparator() {
-        super(BlockKey.class, true);
-    }
-
-    @Override
-    public int compare(WritableComparable a, WritableComparable b) {
-        //return a.compareTo(b);
-        BlockKey keyA = (BlockKey) a;
-        BlockKey keyB = (BlockKey) b;
-        if (keyA.block < keyB.block) return -1;
-        if (keyA.block > keyB.block) return 1;
-        return 0;
+    public TupleKey getKey(Object object) {
+        Person person = (Person) object;
+        return new TupleKey(person.mainInterest(), person.accountId());
     }
 }
