+++ src/main/java/ldbc/snb/datagen/hadoop/miscjob/keychanger/RandomKeySetter.java
@@ -33,27 +33,19 @@
  You should have received a copy of the GNU General Public License
  along with this program; if not, write to the Free Software
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.*/
-package ldbc.snb.datagen.hadoop;
+package ldbc.snb.datagen.hadoop.miscjob.keychanger;
 
-import org.apache.hadoop.io.WritableComparable;
-import org.apache.hadoop.io.WritableComparator;
+import ldbc.snb.datagen.hadoop.key.TupleKey;
+import ldbc.snb.datagen.objects.Person;
 
 /**
- * Created by aprat on 11/9/14.
+ * Created by aprat on 11/17/14.
  */
-public class BlockKeyComparator extends WritableComparator {
+public class RandomKeySetter implements HadoopFileKeyChanger.KeySetter<TupleKey> {
 
-    protected BlockKeyComparator() {
-        super(BlockKey.class, true);
-    }
-
-    @Override
-    public int compare(WritableComparable a, WritableComparable b) {
-        //return a.compareTo(b);
-        BlockKey keyA = (BlockKey) a;
-        BlockKey keyB = (BlockKey) b;
-        if (keyA.block < keyB.block) return -1;
-        if (keyA.block > keyB.block) return 1;
-        return keyA.tk.compareTo(keyB.tk);
+    public TupleKey getKey(Object object) {
+        Person person = (Person) object;
+        return new TupleKey(person.randomId(), person.accountId());
     }
 }
+
