+++ src/main/java/ldbc/snb/datagen/hadoop/serializer/HadoopUpdateStreamSerializer.java
@@ -33,11 +33,12 @@
  You should have received a copy of the GNU General Public License
  along with this program; if not, write to the Free Software
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.*/
-package ldbc.snb.datagen.hadoop;
+package ldbc.snb.datagen.hadoop.serializer;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -48,50 +49,52 @@ import org.apache.hadoop.mapreduce.lib.output.SequenceFileOutputFormat;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.util.List;
 import java.util.zip.GZIPOutputStream;
 
 /**
  * Created by aprat on 10/15/14.
  */
-public class HadoopUpdateStreamSorterAndSerializer {
+public class HadoopUpdateStreamSerializer {
 
     private Configuration conf;
 
-    public static class HadoopUpdateStreamSorterAndSerializerReducer extends Reducer<UpdateEventKey, Text, UpdateEventKey, Text> {
+    public static class HadoopUpdateStreamSerializerReducer extends Reducer<LongWritable, Text, LongWritable, Text> {
 
-        private boolean compressed = false;
-        private Configuration conf;
-        private String streamType;
+        private OutputStream out;
 
         protected void setup(Context context) {
-            conf = context.getConfiguration();
-            streamType = conf.get("streamType");
+            Configuration conf = context.getConfiguration();
+            int reducerId = Integer.parseInt(conf.get("reducerId"));
+            int partitionId = Integer.parseInt(conf.get("partitionId"));
+            String streamType = conf.get("streamType");
             try {
-                compressed = Boolean.parseBoolean(conf.get("ldbc.snb.datagen.serializer.compressed"));
+                FileSystem fs = FileSystem.get(conf);
+                if (Boolean.parseBoolean(conf.get("ldbc.snb.datagen.serializer.compressed"))) {
+                    Path outFile = new Path(context.getConfiguration()
+                                                   .get("ldbc.snb.datagen.serializer.socialNetworkDir") + "/updateStream_" + reducerId + "_" + partitionId + "_" + streamType + ".csv.gz");
+                    out = new GZIPOutputStream(fs.create(outFile));
+                } else {
+                    Path outFile = new Path(context.getConfiguration()
+                                                   .get("ldbc.snb.datagen.serializer.socialNetworkDir") + "/updateStream_" + reducerId + "_" + partitionId + "_" + streamType + ".csv");
+                    out = fs.create(outFile);
+                }
             } catch (Exception e) {
                 System.err.println(e.getMessage());
+                e.printStackTrace();
             }
         }
 
         @Override
-        public void reduce(UpdateEventKey key, Iterable<Text> valueSet, Context context)
+        public void reduce(LongWritable key, Iterable<Text> valueSet, Context context)
                 throws IOException, InterruptedException {
-            OutputStream out;
+            for (Text t : valueSet) {
+                out.write(t.toString().getBytes("UTF8"));
+            }
+
+        }
+
+        protected void cleanup(Context context) {
             try {
-                FileSystem fs = FileSystem.get(conf);
-                if (compressed) {
-                    Path outFile = new Path(context.getConfiguration()
-                                                   .get("ldbc.snb.datagen.serializer.socialNetworkDir") + "/updateStream_" + key.reducerId + "_" + key.partition + "_" + streamType + ".csv.gz");
-                    out = new GZIPOutputStream(fs.create(outFile));
-                } else {
-                    Path outFile = new Path(context.getConfiguration()
-                                                   .get("ldbc.snb.datagen.serializer.socialNetworkDir") + "/updateStream_" + key.reducerId + "_" + key.partition + "_" + streamType + ".csv");
-                    out = fs.create(outFile);
-                }
-                for (Text t : valueSet) {
-                    out.write(t.toString().getBytes("UTF8"));
-                }
                 out.close();
             } catch (Exception e) {
                 System.err.println(e.getMessage());
@@ -99,33 +102,28 @@ public class HadoopUpdateStreamSorterAndSerializer {
         }
     }
 
-
-    public HadoopUpdateStreamSorterAndSerializer(Configuration conf) {
+    public HadoopUpdateStreamSerializer(Configuration conf) {
         this.conf = new Configuration(conf);
     }
 
-    public void run(List<String> inputFileNames, String type) throws Exception {
+    public void run(String inputFileName, int reducer, int partition, String type) throws Exception {
 
-        int numThreads = conf.getInt("ldbc.snb.datagen.generator.numThreads", 1);
+        conf.setInt("reducerId", reducer);
+        conf.setInt("partitionId", partition);
         conf.set("streamType", type);
 
         Job job = Job.getInstance(conf, "Update Stream Serializer");
-        job.setMapOutputKeyClass(UpdateEventKey.class);
+        job.setMapOutputKeyClass(LongWritable.class);
         job.setMapOutputValueClass(Text.class);
-        job.setOutputKeyClass(UpdateEventKey.class);
+        job.setOutputKeyClass(LongWritable.class);
         job.setOutputValueClass(Text.class);
-        job.setJarByClass(HadoopUpdateStreamSorterAndSerializerReducer.class);
-        job.setReducerClass(HadoopUpdateStreamSorterAndSerializerReducer.class);
-        job.setNumReduceTasks(numThreads);
+        job.setJarByClass(HadoopUpdateStreamSerializerReducer.class);
+        job.setReducerClass(HadoopUpdateStreamSerializerReducer.class);
+        job.setNumReduceTasks(1);
         job.setInputFormatClass(SequenceFileInputFormat.class);
         job.setOutputFormatClass(SequenceFileOutputFormat.class);
-        job.setPartitionerClass(HadoopUpdateEventKeyPartitioner.class);
-        job.setGroupingComparatorClass(UpdateEventKeyGroupComparator.class);
-        //job.setSortComparatorClass(UpdateEventKeySortComparator.class);
 
-        for (String s : inputFileNames) {
-            FileInputFormat.addInputPath(job, new Path(s));
-        }
+        FileInputFormat.setInputPaths(job, new Path(inputFileName));
         FileOutputFormat.setOutputPath(job, new Path(conf.get("ldbc.snb.datagen.serializer.hadoopDir") + "/aux"));
         if (!job.waitForCompletion(true)) {
             throw new Exception();
